+++ lib/rails/auth/config_builder.rb
@@ -31,22 +31,15 @@ module Rails
       def production(
         config,
         cert_filters: nil,
-        require_cert: false,
-        ca_file: nil,
         error_page: Rails.root.join("public/403.html"),
         monitor: nil
       )
-        raise ArgumentError, "no cert_filters given but require_cert is true" if require_cert && !cert_filters
-        raise ArgumentError, "no ca_file given but cert_filters were set"     if cert_filters && !ca_file
-
         error_page_middleware(config, error_page)
 
         if cert_filters
           config.middleware.insert_before Rails::Auth::ACL::Middleware,
                                           Rails::Auth::X509::Middleware,
-                                          require_cert: require_cert,
                                           cert_filters: cert_filters,
-                                          ca_file:      ca_file,
                                           logger:       Rails.logger
         end